requirements.txt
### requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-tavily
langchain-community>=0.3.0
langchain_huggingface
langchain-openai
langgraph>=0.2.0

fastapi
fastapi-sessions
fastapi-redis-session
uvicorn[standard]

python-jose 
passlib
regex
bcrypt
python-multipart
python-dotenv

pymongo

email-validator>=2.0.0
pydantic>=2.0.0

sentence-transformers

google-auth 
google-auth-oauthlib 
google-auth-httplib2 
google-api-python-client

itsdangerous 

minio